+++ src/Extensions/ZygoteExtensions.jl
@@ -123,7 +123,6 @@ function val_and_jacobian(f, args...;ignore_pullback=[])
   (val=y, grad=out)
 end
 
-
 # Equivalent `Zygote.withgradient` call that is compatible with PartitionedArrays
 """
     val_and_gradient(f, args...)
@@ -135,3 +134,16 @@ function val_and_gradient(f, args...)
   grad = back(Zygote.sensitivity(y))
   (val=y, grad)
 end
+
+"""
+    Hvp(f,p,v)
+
+Compute the Hessian-vector product of f at p in the direction of v using forward-over-reverse AD.
+"""
+function Hvp(f,p,v)
+  ∇f = p->val_and_gradient(f,p).grad[1]
+  ForwardDiff.derivative(α -> ∇f(p + α*v), 0)
+end
+
+# FowardDiff extensions
+@inline ForwardDiff.extract_derivative(::Type{T}, y::PVector) where {T} = Base.Fix1(ForwardDiff.extract_derivative,T).(y)