+++ test/mpi/StateMapTests/AffineFEStateMapTest.jl
@@ -101,7 +101,7 @@ function main(distribute,mesh_partition)
   end
 end
 
-with_mpi() do distribute
+with_debug() do distribute
   main(distribute,(2,2))
 end