+++ test/seq/StateMapTests/runtests.jl
@@ -21,6 +21,9 @@ using Test
   include("Transient_CrankNicolson_AffineFEStateMap.jl")
   include("Transient_CrankNicolson_ConstantMats_AffineFEStateMap.jl")
   include("Transient_CrankNicolson_NonlinearFEStateMap.jl")
+  # Second order
+  include("SecondOrderTests.jl")
+  include("SecondOrderTestsMPI_seq.jl")
 end
 
 end # module